+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 # Version 0.0.7
 
-- Rust client examples in programs/multisig-lite/examples
-- Run Rust client examples in CI
-- Run anchor test in CI
+- Rust client examples in programs/multisig-lite/examples.
+- Run Rust client examples in CI.
+- Run anchor test in CI.
 - Add preliminary solana-program-test based functional tests.
 
 # Version 0.0.6